refactor(domain): resolveBust 메서드를 Participant로 이동
